Cleanup Fake backend code
